chore: reorganize folder structure for better code organization
- Create feature-specific folder structure for help, auth, questions, statements, and wizard
- Move UI component files to their corresponding feature folders
- Create index.ts barrel files for cleaner imports
- Update import paths in affected files
- Fix TypeScript interface definitions in Dropdown component
- Ensure proper file organization following feature-based architecture
