+++ src/components/layout/index.ts
@@ -1,0 +1,10 @@
+/**
+ * Layout Components Module
+ * 
+ * Provides the main structural components for page layout,
+ * including the overall app container, header, and footer.
+ */
+
+export { default as Header } from './Header';
+export { default as Footer } from './Footer';
+export { default as MainPage } from './MainPage';
