+++ src/components/modals/index.ts
@@ -1,0 +1,12 @@
+/**
+ * Modal Components Module
+ * 
+ * Provides reusable modal dialog components for various purposes
+ * throughout the application.
+ */
+
+export { default as GratitudeModal } from './GratitudeModal';
+export { default as PrivacyModal } from './PrivacyModal';
+export { default as ShareEmailModal } from './ShareEmailModal';
+export { default as TermsModal } from './TermsModal';
+export { default as UserDataModal } from './UserDataModal';
