+++ src/components/ui/index.ts
@@ -1,0 +1,44 @@
+/**
+ * UI Components Module
+ * 
+ * Provides reusable UI components that can be used across the application.
+ * These components are the building blocks for more complex components.
+ */
+
+// Export individual UI components
+export { Button } from './Button';
+export { Input } from './Input';
+export { Slot } from './Slot';
+export { ConfirmationDialog } from './ConfirmationDialog';
+
+// Export dialog components
+export {
+  SimpleDialog,
+  SimpleDialogContent,
+  SimpleDialogTitle,
+  SimpleDialogDescription,
+  SimpleDialogFooter,
+  SimpleDialogHeader
+} from './Dialog';
+
+// Export dropdown components
+export {
+  SimpleDropdownMenu,
+  SimpleDropdownMenuTrigger,
+  SimpleDropdownMenuContent,
+  SimpleDropdownMenuItem,
+  SimpleDropdownMenuSeparator
+} from './Dropdown';
+
+// Export tooltip components
+export {
+  Tooltip,
+  TooltipTrigger,
+  TooltipContent,
+  TooltipProvider
+} from './BetterTooltip';
+
+// Export question counter components
+export { default as QuestionCounter } from './questionCounter/QuestionCounter';
+export { default as LargeCircularQuestionCounter } from './questionCounter/LargeCircularQuestionCounter';
+export { default as SmallCircularQuestionCounter } from './questionCounter/SmallCircularQuestionCounter';
