+++ src/features/help/components/WelcomePanel.tsx
