+++ src/features/help/components/index.ts
@@ -1,0 +1,3 @@
+export { default as HelpButton } from './HelpButton';
+export { default as HelpCenter } from './HelpCenter';
+export { default as WelcomePanel } from './WelcomePanel';
