+++ src/features/questions/hooks/useQuestions.ts
@@ -1,5 +1,5 @@
 import { useContext } from 'react';
-import { QuestionsContext } from '../../../providers/QuestionsContext';
+import { QuestionsContext } from '../context/QuestionsContext';
 
 export function useQuestions() {
   const context = useContext(QuestionsContext);
