+++ src/features/statements/components/index.ts
@@ -1,0 +1,7 @@
+// Barrel file for statements components
+export { default as ActionForm } from './ActionForm';
+export { default as ActionLine } from './ActionLine';
+export { default as ActionsCounter } from './ActionsCounter';
+export { default as QuestionCard } from './QuestionCard';
+export { default as StatementItem } from './StatementItem';
+export { default as StatementList } from './StatementList';
