+++ src/features/statements/index.ts
@@ -1,0 +1,3 @@
+// Barrel file for statements feature exports
+export { useEntries } from './hooks/useEntries';
+export { EntriesProvider } from './context/EntriesProvider';
