+++ src/features/wizard/components/index.ts
@@ -1,0 +1,10 @@
+// Barrel file for wizard components
+export { default as StatementWizard } from './StatementWizard';
+export { default as StatementPreview } from './StatementPreview';
+export { EditStatementModal } from './EditStatementModal';
+export { default as FilterBar } from './FilterBar';
+export { default as SentimentVerbPicker } from './SentimentVerbPicker';
+export { default as StepContainer } from './StepContainer';
+export { default as SubjectTiles } from './SubjectTiles';
+export { default as VerbGrid } from './VerbGrid';
+export { default as VerbSelector } from './VerbSelector';
