+++ src/features/wizard/components/steps/index.ts
@@ -1,0 +1,7 @@
+// Barrel file for wizard step components
+export { CategoryStep } from './CategoryStep';
+export { ComplementStep } from './ComplementStep';
+export { ObjectStep } from './ObjectStep';
+export { PrivacyStep } from './PrivacyStep';
+export { SubjectStep } from './SubjectStep';
+export { VerbStep } from './VerbStep';
