+++ src/features/wizard/index.ts
@@ -1,0 +1,2 @@
+// Barrel file for wizard feature exports
+export { default as StatementWizard } from './components/StatementWizard';
